Remove the web services to init modules/tenants base. Params to MainVerticle
Using inmemory "database" for the unit tests, so we don't need the
initTenants and initModules web services - those were too dangerous
to leave around, would delete whole databases! This also needed
a small fix in the way parameters are passed to the okapi main
verticle - now it tries System.getProperty(), config.getString(),
and a hard-coded default, in that order.