Close down vert.x in the sample/auth modules
Towards #63 - this releases the listening port immediately, so we
are less likely to get the "Address already in use". If we wait until
the OS kills the process and collects loose ports, it may seem to be
busy for longer, overlapping with the attempt to open the same port
by the new incarnation of the module.