Setting mongo_db_init=1 resets database during start
Magic mongo_transient_test has no special semantics anymore.
Allows me to make a test from a clean slate with a private
database: that is 1st run and 2nd run is same database, but
only 1st run would reset it. I was impossible to test the behavior
of issue 81 without this feature. With magic database name it
would reset on all runs.