No slash in Vertx routes for POST/GET of /_/modules
The RAML has no trailing slash. The Unit tests has no trailing slash
so, it begs the question why the tests did succeed before. According
to Vert.x doc the router spec is a prefix .. And /_/modules/ is not
a prefix of /_/modules . With this in place a trailing slash is still
accepted (and if I understand Vert.x doc correctely also other things).
A trailing slash is not proper WRT RAML. Relates to #82.