Allow node to be updated in Hazelcast map. Issue #127
Unfortunately when we undeploy a node - the addShutdownHook
code is executed after the Hazelcast message is gone.. Hence
when we undeploy, the discovery is not informed. So when the node
is deployed again, it will be already present. The node "host" system
property is by default localhost. That's not adequate for a multi node
where they should be unique (just as their URLs must be unique, to be
able to address them!):