FOLIO-221: Clean up after each test
Make sure we always leave the database in a clean state, so the
next test may assume its data will not collide with old test data.
This has hurt us in strange ways, since the tests are not run in
a deterministic order.

Also added a check to verify the db is clean at the end of every test.