Implement Okapi-238: System level interfaces
When enabling and disabling modules, checks if the module supports
an interface called "_tenant" (note the underscore), and if so,
makes a request to the "well-known" path of "/_/tenant".

If the moduleDescriptor has a tenantInterface field, this takes precedence,
but produces a warning in the log that this is deprecated. It will be removed
later.