First rough version check in place
When starting Okapi checks the version that is about to start against
the one we have in the database (if any). If there is a newer version
in the db, Okapi refuses to start - you should be using the latest. If
the starting Okapi is newer than the database, it will update the
version that is enabled for the supertenant.

At the moment uses trivial string compare for the version check.
That works for 1.8 and 1.9, but will fail at 1.10 - Adam is working
on refactoring the version checks into a module that will be used
here too.