Translate '?' to '-' only if -q option specified
It turns out that the Markdown-Viewer Chrome extension that I use
locally expects a different format in its heading links from what the
GitHub Markdown viewer uses. I've made it so that the default
behaviour is as before: question marks in a heading are simply
discarded from the link that gets generated -- so that things work
right when viewed on GitHub. But if the new '-q' option is specified,
then question marks are translated into hyphens -- so that the links
work in the Chrome extension.