Two new properties metadata and env in ModuleDescriptor OKAPI-732
The env property is supposed to be a complete list of the environment
variables that the module supports. These should be passed verbatim
for the module if in a container (regardless of deployment). For
everything, there is new matadata property that can be used to further
describe to module (opaque to Okapi itself).