Metadata and env in md (#798) OKAPI-732
* Two new properties metadata and env in ModuleDescriptor OKAPI-732

The env property is supposed to be a complete list of the environment
variables that the module supports. These should be passed verbatim
for the module if in a container (regardless of deployment). For
everything, there is new matadata property that can be used to further
describe to module (opaque to Okapi itself).

Unit test fix: test pre filter content after some wait

Cluster fails in Jenkins sometimes. So change MultiTenantTest to use mode dev instead.