Allow module and interface discovery for same tenant OKAPI-863
Okapi on its own can not check if tenant is valid, so it might
pass no required permissions to auth. But auth will still fail
if tenant in token doesn't match. Internally this is done in
the routing entry of Okapi module to use a new property which
lists permissions to use when tenant matches {tenantId} in
pathPattern. This solution is, thus, backwards compatible with
OKAPI-835 for a tenant match.