Allow module and interface discovery for same tenant OKAPI-863 (#985)
New ModuleDescriptor RoutingEntry property permissionsRequiredTenant with the required permissions
if tenant matches that of tenantId pattern in pathPattern. Okapi can not check if tenant is valid,
so it might pass permissionsRequiredTenant to auth. However, the token is validated by mod-authtoken
and it will be rejected if tenantId is not identical to that in the token.