OKAPI-971: Fail unit tests if docker is not available (needed for testcontainers)
FOLIO switches from embedded-postgresql to testcontainers.org that use
Docker PostgreSQL containers for tests in RMB based modules, RMB and Okapi:
FOLIO-1845

It no longer makes sense to skip tests if Docker is not available.

Tests should fail by default if we cannot run all tests so that any Docker
malfunction on the developer machine and on CI is noticed immediately. It
is misleading if such tests are skipped by default, see comments on OKAPI-970.

DockerModuleHandleTest, ModuleTest and PostgresHandleTest require Docker.

This pull request removes

    @Testcontainers(disabledWithoutDocker = true)

PostgresHandleTest. (The other two files already fail on missing Docker).
And it adjust README.

If there really is a need to run only those tests that don't require
Docker please create a new Jira for this feature. This might be
implemented by adding a `withoutDocker` maven profile to the pom.xml
that can be used in this way:

    mvn clean install -PwithoutDocker