Use HTTP check for module readiness in deployment OKAPI-984 (#1084)
Wait with HTTP check for deployment OKAPI-984

Use WebClient on module host+port to see if module is alive, rather than just TCP socket against host and port. The latter is unreliable in cases of Containers where exposed port is ready "immediately". This in turn fixes Okapi startup with mod-permissions being called for the purpose of upgrading the okapi module for a tenant.

Co-authored-by: julianladisch <julianladisch@users.noreply.github.com>