Inform mod-permissions when module is disabled OKAPI-982 (#1082)
When a module is disabled call tenant permissions with the module being disabled and an empty list.
This in turn allows mod-permissions to deprecate or remove the permissions as they are no longer defined
for the tenant - provided that mod-permissions keep track of which module defines a permission.

Install parameter 'invoke' extended. It may be a pattern that specifies which modules to invoke during
install, eg invoke=mod-permissions.* ,would specify that only mod-permissions should be called. 

New configuration/property  'enable_system_auth'  which is true by default. If false, even
if auth is enabled it, the auth module is not be invoked for system calls. This should is for testing purposes
only - never for production.