OKAPI-1025: /_/proxy/tenants/{tenant_id}/upgrade with body
/_/proxy/tenants/\{tenant_id}/upgrade upgrades each module to the latest version
and therefore doesn't needs a body to list the version to upgrade to.

Steps to Reproduce:

* POST /_/proxy/tenants/\{tenant_id}/upgrade with body

Expected Results:

Error message about the useless body.

Actual Results:

Upgrades to latest versions ignoring the version specified in the body.

Additional Information:

The documentation in
https://github.com/folio-org/okapi/blob/v4.8.2/okapi-core/src/main/raml/okapi.raml
correctly explains that no body is needed and it upgrades to the latest version.
(The wording can be improved, though.)

Okapi provides
 * /_/proxy/tenants/\{tenant_id}/install
 * /_/proxy/tenants/\{tenant_id}/upgrade

being misleading at first glance because one might assume that install is intended
for initial install and upgrade is intended for upgrades.
Rejecting upgrade with body will prevent any unintentional results.