Add cleanup parameter removeDependencies
When true, the cleanup process will happily remove all modules that
have broken dependencies as a result of the clean up process.
Unfortnately, there are modules out there that depend on interface
only available in snapshots. When this parameter is false (default),
the broken dependencies are reported and operation is aborted.