Further tests of OKAPI-1037
test auth module modified a bit to behave more like
mod-authtoken. This, in turn, enables cahcing a bit more -
since caching requires X-Okapi-Module-Tokens to be present.
This revaled another fault with token caching with routing
entries including permissionsRequiredTenant.

A test for use of X-Okapi-Addtional-Token was removed as caching
makes these fail. And for good reason: this header is not used
anywhere but in Okpai and its test auth module. OKAPI-656.