Only put in token in cache if auth returns OK OKAPI-1037 (#1148)
Only put in token in cache if auth returns OK OKAPI-1037

test auth module modified a bit to behave more like
mod-authtoken. This, in turn, enables caching a bit more -
since caching requires X-Okapi-Module-Tokens to be present.
This revealed another fault with token caching with routing
entries including permissionsRequiredTenant.

A test for use of X-Okapi-Additonal-Token was removed as caching
makes these fail. And for good reason: this header is not used
anywhere but in Okapi and its test auth module. OKAPI-656.