install with action=enable and purge OKAPI-1064 (#1186)
Enable with purge will make two calls to the module to be enabled.
The first being disable (as if the module was disabled). Errors
are ignored for this first operation as some modules returns 400
if a module was not enabled already. The second will proceed as
usual. This allows to force a clean slate for a module.