Too much to mention
OK. Three types of clients: the login client, a yet-to-complete
fresh token client that takes refresh token and gest access tokens
from that as necessary. And finally, the API client which is
complete and so simple that it hurts.

TokenCache moved into generic class Expiry, so that we can have
a cache for login based client and another for refresh token
based operation. The API token based one, obviously does not
need this!

Using remove + put to ensure updated entry is put it in front of
linked list in LinkedHashMap.