OKAPI-1142: HTTP idle timeout
Option to configure an HTTP idle timeout for Okapi.

Netty and Vert.x disable it by default.

If enabled Okapi can log the API that is still open after a long time. This helps to investigate
https://issues.folio.org/browse/OKAPI-1141

Approach:

HttpClientOptions.setIdleTimeout and configuration option in conf file and as command line parameter.