OKAPI-1207: LoginClient/RefreshClient: Suppress stacktrace, compress JSON
https://folio-org.atlassian.net/browse/OKAPI-1207

okapi-common' LoginClient and RefreshClient fail if user is blocked or password is wrong.
It causes a log entry with stacktrace and multi-line JSON body.

The stacktrace is useless because it only shows Netty and Vert.x interals.

The JSON body is pretty-printed taking multiple lines and should be compressed into a single line.

The useless verbosity blows up the logs and makes it more difficult for sysops to browse the logs.