OKAPI-1223: Declare multiple interface as optional dependency for Eureka
https://folio-org.atlassian.net/browse/OKAPI-1223

Craig McNally writes:

When sidecars initialize they make a call to mgr-applications to get information about the module they're associated with.  This includes information about which other modules the associated module might call.  The declared interface dependencies in the module descriptors is used for this purpose (both required and optional).  So for example, mod-bulk-operations declares an optional dependency on the custom-fields interface.  At startup, the mod-bulk-operations sidecar makes a call to GET /modules/mod-bulk-operations-2.2.6 and gets back information about mod-bulk-operations (for handling incoming/ingress requests), but also for other modules which mod-bulk-operations calls, including any module which provides the custom-fields interface.

If these interface dependencies are not declared, the sidecar will not know how to route a module-to-module call.  What happens next is configurable in the sidecar... the sidecar could return an error response, or instead route the call to kong.  This will usually work, but comes with potential performance implications, and could make troubleshooting more difficult since I believe kong will assign the request a new request Id, causing problems when tracing a request through the system.

So, I suggest we declare dependencies on interfaces w/ type: multiple as optional dependencies.